+++ expenser-server/internal/handler/handleTransaction.go
@@ -29,11 +29,11 @@ func HandleTransactionGet(data model.Config) http.HandlerFunc {
 
 func HandleTransactionCreate(data model.Config) http.HandlerFunc {
 	type parameters struct {
-		Name   string `json:"name"`
-		Amount int    `json:"amount"`
-		Type   string `json:"type"`
-		Date   string `json:"date"`
-		Note   string `json:"note"`
+		Name   string  `json:"name"`
+		Amount float64 `json:"amount"`
+		Type   string  `json:"type"`
+		Date   string  `json:"date"`
+		Note   string  `json:"note"`
 	}
 
 	type response struct {
@@ -66,11 +66,11 @@ func HandleTransactionCreate(data model.Config) http.HandlerFunc {
 
 func HandleTransactionUpdate(data model.Config) http.HandlerFunc {
 	type parameters struct {
-		Name   string `json:"name"`
-		Amount int    `json:"amount"`
-		Type   string `json:"type"`
-		Date   string `json:"date"`
-		Note   string `json:"note"`
+		Name   string  `json:"name"`
+		Amount float64 `json:"amount"`
+		Type   string  `json:"type"`
+		Date   string  `json:"date"`
+		Note   string  `json:"note"`
 	}
 
 	type response struct {
